backend/requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.0.3
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.3.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gunicorn==20.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.0
requests==2.32.3
rsa==4.9
urllib3==2.2.1
Werkzeug==2.1.2